ai-chatapp/requirements.txt
streamlit==1.40.2
openai==1.57.3
python-dotenv==1.0.1